Move BC permission utils out of enterprise
