+++ client/web/src/enterprise/organizations/settings/sidebaritems.ts
@@ -1,6 +1,6 @@
+import { canWriteBatchChanges } from '../../../batches/utils'
 import { OrgSettingsSidebarItems } from '../../../org/settings/OrgSettingsSidebar'
 import { orgSettingsSideBarItems } from '../../../org/settings/sidebaritems'
-import { canWriteBatchChanges } from '../../batches/utils'
 
 export const enterpriseOrgSettingsSideBarItems: OrgSettingsSidebarItems = [
     ...orgSettingsSideBarItems,
