+++ client/web/src/enterprise/user/settings/routes.tsx
@@ -1,8 +1,8 @@
 import { lazyComponent } from '@sourcegraph/shared/src/util/lazyComponent'
 
+import { canWriteBatchChanges } from '../../../batches/utils'
 import { userSettingsAreaRoutes } from '../../../user/settings/routes'
 import { UserSettingsAreaRoute } from '../../../user/settings/UserSettingsArea'
-import { canWriteBatchChanges } from '../../batches/utils'
 import { SHOW_BUSINESS_FEATURES } from '../../dotcom/productSubscriptions/features'
 import type { ExecutorsUserAreaProps } from '../../executors/ExecutorsUserArea'
